App inherits Blueprint
